+++ src/graphics/text/rc.rs
@@ -51,7 +51,7 @@ pub struct Text {
     #[doc(hidden)]
     text: *mut ffi::sfText,
     #[doc(hidden)]
-    string_length: uint,
+    string_length: usize,
     #[doc(hidden)]
     font: Option<Rc<RefCell<Font>>>
 }
@@ -85,7 +85,7 @@ impl Text {
     /// Return Some(Text) or None
     pub fn new_init(string: &str,
                     font: Rc<RefCell<Font>>,
-                    character_size: uint) ->Option<Text> {
+                    character_size: usize) ->Option<Text> {
         let text = unsafe { ffi::sfText_create() };
         if text.is_null() {
             None
@@ -167,9 +167,9 @@ impl Text {
     /// Get the size of the characters
     ///
     /// Return the size of the characters
-    pub fn get_character_size(&self) -> uint {
+    pub fn get_character_size(&self) -> usize {
         unsafe {
-            ffi::sfText_getCharacterSize(self.text) as uint
+            ffi::sfText_getCharacterSize(self.text) as usize
         }
     }
 
@@ -248,7 +248,7 @@ impl Text {
     ///
     /// # Arguments
     /// * size - The new character size, in pixels
-    pub fn set_character_size(&mut self, size: uint) -> () {
+    pub fn set_character_size(&mut self, size: usize) -> () {
         unsafe {
             ffi::sfText_setCharacterSize(self.text, size as c_uint)
         }
