+++ src/network/http.rs
@@ -37,11 +37,11 @@ use ffi::network::http as ffi;
 #[derive(Clone, PartialEq, Eq, PartialOrd, Ord, Show, Copy)]
 pub enum Method {
     /// Request in get mode, standard method to retrieve a page
-    Get = ffi::GET as int,
+    Get = ffi::GET as isize,
     /// Request in post mode, usually to send data to a page
-    Post = ffi::POST as int,
+    Post = ffi::POST as isize,
     /// Request a page's header only
-    Head = ffi::HEAD as int
+    Head = ffi::HEAD as isize
 }
 
 /// Status code returned by a serveur.
@@ -49,59 +49,59 @@ pub enum Method {
 pub enum Status {
     // 2xx: success
     /// Most common code returned when operation was successful
-    Ok                  = ffi::OK as int,
+    Ok                  = ffi::OK as isize,
     /// The resource has successfully been created
-    Created             = ffi::CREATED as int,
+    Created             = ffi::CREATED as isize,
     /// The request has been accepted, but will be processed later by the server
-    Accepted            = ffi::ACCEPTED as int,
+    Accepted            = ffi::ACCEPTED as isize,
     /// Sent when the server didn't send any data in return
-    NoContent           = ffi::NOCONTENT as int,
+    NoContent           = ffi::NOCONTENT as isize,
     /// The server informs the client that it should clear the view (form) that caused the request to be sent
-    ResetContent        = ffi::RESETCONTENT as int,
+    ResetContent        = ffi::RESETCONTENT as isize,
     /// The server has sent a part of the resource, as a response to a partial GET request
-    PartialContent      = ffi::PARTIALCONTENT as int,
+    PartialContent      = ffi::PARTIALCONTENT as isize,
 
     // 3xx: redirection
     /// The requested page can be accessed from several locations
-    MultipleChoices     = ffi::MULTIPLECHOICES as int,
+    MultipleChoices     = ffi::MULTIPLECHOICES as isize,
     /// The requested page has permanently moved to a new location
-    MovedPermanently    = ffi::MOVEDPERMANENTLY as int,
+    MovedPermanently    = ffi::MOVEDPERMANENTLY as isize,
     /// The requested page has temporarily moved to a new location
-    MovedTemporarily    = ffi::MOVEDTEMPORARILY as int,
+    MovedTemporarily    = ffi::MOVEDTEMPORARILY as isize,
     /// For conditionnal requests, means the requested page hasn't changed and doesn't need to be refreshed
-    NotModified         = ffi::NOTMODIFIED as int,
+    NotModified         = ffi::NOTMODIFIED as isize,
 
     // 4xx: client error
     /// The server couldn't understand the request (syntax error)
-    BadRequest          = ffi::BADREQUEST as int,
+    BadRequest          = ffi::BADREQUEST as isize,
     /// The requested page needs an authentification to be accessed
-    Unauthorized        = ffi::UNAUTHORIZED as int,
+    Unauthorized        = ffi::UNAUTHORIZED as isize,
     /// The requested page cannot be accessed at all, even with authentification
-    Forbidden           = ffi::FORBIDDEN as int,
+    Forbidden           = ffi::FORBIDDEN as isize,
     /// The requested page doesn't exist
-    NotFound            = ffi::NOTFOUND as int,
+    NotFound            = ffi::NOTFOUND as isize,
     /// The server can't satisfy the partial GET request (with a "Range" header field)
-    RangeNotSatisfiable = ffi::RANGENOTSATISFIABLE as int,
+    RangeNotSatisfiable = ffi::RANGENOTSATISFIABLE as isize,
 
     // 5xx: server error
     /// The server encountered an unexpected error
-    InternalServerError = ffi::INTERNALSERVERERROR as int,
+    InternalServerError = ffi::INTERNALSERVERERROR as isize,
     /// The server doesn't implement a requested feature
-    NotImplemented      = ffi::NOTIMPLEMENTED as int,
+    NotImplemented      = ffi::NOTIMPLEMENTED as isize,
     /// The gateway server has received an error from the source server
-    BadGateway          = ffi::BADGATEWAY as int,
+    BadGateway          = ffi::BADGATEWAY as isize,
     /// The server is temporarily unavailable (overloaded, in maintenance, ...)
-    ServiceNotAvailable = ffi::SERVICENOTAVAILABLE as int,
+    ServiceNotAvailable = ffi::SERVICENOTAVAILABLE as isize,
     /// The gateway server couldn't receive a response from the source server
-    GatewayTimeout      = ffi::GATEWAYTIMEOUT as int,
+    GatewayTimeout      = ffi::GATEWAYTIMEOUT as isize,
     /// The server doesn't support the requested HTTP version
-    VersionNotSupported = ffi::VERSIONNOTSUPPORTED as int,
+    VersionNotSupported = ffi::VERSIONNOTSUPPORTED as isize,
 
     // 10xx: SFML custom codes
     /// Response is not a valid HTTP one
-    InvalidResponse     = ffi::INVALIDRESPONSE as int,
+    InvalidResponse     = ffi::INVALIDRESPONSE as isize,
     /// Connection with server failed
-    ConnectionFailed    = ffi::CONNECTIONFAILED as int
+    ConnectionFailed    = ffi::CONNECTIONFAILED as isize
 }
 
 /// Encapsulation of an HTTP request
